refactoring, go fmt
